+++ src/F1.Infrastructure/Data/Entities/MigrationImportCalculatedTotalEntity.cs
@@ -5,5 +5,5 @@ public sealed class MigrationImportCalculatedTotalEntity
     public long Id { get; set; }
     public Guid ImportRunId { get; set; }
     public string Subject { get; set; } = string.Empty;
-    public int CalculatedTotalPoints { get; set; }
+    public decimal CalculatedTotalPoints { get; set; }
 }
